backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
Pillow==10.1.0
django-filter==23.4
PyJWT==2.8.0
python-decouple==3.8
requests==2.31.0
django-extensions==3.2.3
factory-boy==3.3.0
Faker==20.1.0